Update AppCourses to handle both images and descriptions
defaultCards.map() loops over each of the 12 default cards.
card refers to the current default card. index refers to the position of the card in the array. courseStore.images[index] checks if there is an image available at the current index in the courseStore.images array; If there is an image at that index, it uses the fetched image and if there isn’t an image, it falls back to the default card image, which is just an empty string "". The same thing applies to the description.